# mksprite: compute YUV error for IHQ