@kannoneer 2023-09-17
The original images for reference. 
![grass](https://github.com/DragonMinded/libdragon/assets/99533466/c7b83b1c-3336-4195-9d8f-959ee75b3ef6) ![circuit](https://github.com/DragonMinded/libdragon/assets/99533466/0fa7fce5-65da-4e61-8674-1362a70c21f1) ![stone](https://github.com/DragonMinded/libdragon/assets/99533466/e2fdd93e-f3c2-44cb-966d-fb4f0397f7c6) ![castle](https://github.com/DragonMinded/libdragon/assets/99533466/8ab6cba0-e315-42a7-851f-5cc46df5a62b)