@rasky 2023-09-17
I think the glitches might be a bug in my bilinear interpolation, I’ll have a look. Patch is ok for me